#366 - Refactor view package
- Move pageable behavior command handling to layoutable behavior
- Add locatable trait to provide getUrl() and getRoute() methods
- Add modellable trait to provide model and entity specific getters
- Remove pageable behavior
- Cleanup view classes
