+++ code/site/components/com_pages/view/csv.php
@@ -9,35 +9,7 @@
 
 class ComPagesViewCsv extends KViewCsv
 {
-    public function isCollection()
-    {
-        return (bool) !$this->getModel()->getState()->isUnique();
-    }
-
-    public function getRoute($page = null, $query = array(), $escape = false)
-    {
-        return $this->getBehavior('routable')->getRoute($page, $query, $escape);
-    }
-
-    public function getUrl($url = null)
-    {
-        if(!empty($url))
-        {
-            if($url instanceof KHttpUrlInterface)
-            {
-                $result = clone $url;
-                $result->setUrl(parent::getUrl()->toString(KHttpUrl::AUTHORITY));
-            }
-            else
-            {
-                $result = clone parent::getUrl();;
-                $result->setUrl($url);
-            }
-        }
-        else $result = parent::getUrl();
-
-        return $result;
-    }
+    use ComPagesViewTraitModellable, ComPagesViewTraitLocatable;
 
     protected function _fetchData(KViewContext $context)
     {
