+++ code/site/components/com_pages/view/html.php
@@ -9,12 +9,21 @@
 
 class ComPagesViewHtml extends ComKoowaViewHtml
 {
+    use ComPagesViewTraitModellable, ComPagesViewTraitLocatable;
+
     protected function _initialize(KObjectConfig $config)
     {
         $config->append([
-            'behaviors'   => ['routable', 'pageable', 'layoutable'],
+            'behaviors'   => ['layoutable'],
             'auto_fetch'  => false,
-            'template_filters' => ['asset', 'meta'],
+            'template_filters'   => ['asset', 'meta'],
+            'template_functions' => [
+                'page'        => [$this, 'getPage'],
+                'collection'  => [$this, 'getCollection'],
+                'state'       => [$this, 'getState'],
+                'direction'   => [$this, 'getDirection'],
+                'language'    => [$this, 'getLanguage'],
+            ],
         ]);
 
         parent::_initialize($config);
@@ -25,51 +34,6 @@ class ComPagesViewHtml extends ComKoowaViewHtml
         $content = $this->getContent();
 
         return trim($content);
-    }
-
-    public function isCollection()
-    {
-        return (bool) !$this->getModel()->getState()->isUnique();
-    }
-
-    public function getTitle()
-    {
-        $title = '';
-        if($page = $this->getModel()->getPage())
-        {
-            $title = $page->title ?: '';
-
-            if($this->getModel()->getState()->isUnique()) {
-                $title = $this->getCollection()->get('title', $title);
-            }
-        }
-
-        return $title;
-    }
-
-    public function getRoute($page = null, $query = array(), $escape = false)
-    {
-        return $this->getBehavior('routable')->getRoute($page, $query, $escape);
-    }
-
-    public function getUrl($url = null)
-    {
-        if(!empty($url))
-        {
-            if($url instanceof KHttpUrlInterface)
-            {
-                $result = clone $url;
-                $result->setUrl(parent::getUrl()->toString(KHttpUrl::AUTHORITY));
-            }
-            else
-            {
-                $result = clone parent::getUrl();;
-                $result->setUrl($url);
-            }
-        }
-        else $result = parent::getUrl();
-
-        return $result;
     }
 
     protected function _fetchData(KViewContext $context)
